+++ src/main/java/com/accenture/granja/repository/GranjaRepository.java
@@ -1,7 +1,10 @@
 package com.accenture.granja.repository;
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.accenture.granja.model.Granja;
-public interface GranjaRepository extends JpaRepository<Granja, Long>{
+public interface GranjaRepository extends JpaRepository<Granja, Integer>{
+
 
 }
